ai_service_py/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
langchain==0.3.7
langchain-mistralai==0.2.6
python-dotenv==1.0.1
